RG-2542 Renamed a CSS file to avoid classnames clash
